requirements.txt
torch==1.8.1
torchvision==0.9.1
pytorch-lightning==1.5.4
PyYAML==6.0
tensorboard==2.7.0
kornia==0.6.2
torch-optimizer==0.3.0
click==8.0.3
matplotlib==3.5.0
cairosvg==2.5.2
svgpathtools==1.4.4
numpy==1.21.4
Pillow==8.4.0
timm==0.4.12
lpips==0.1.4
segmentation-models-pytorch==0.2.1
transformer_encoder==0.0.3
onnx==1.14.0
protobuf==3.20.3
svgelements==1.9.5
pandas==2.0.3
scipy==1.11.1
